- clean architecture implemented in more "canonical" way; - .gitignore updated;
