+++ app/src/main/java/com/gmail/laktionov/lyricsfinder/di/DataModule.kt
@@ -1,14 +1,16 @@
 package com.gmail.laktionov.lyricsfinder.di
 
 import android.content.res.Resources
-import com.gmail.laktionov.lyricsfinder.domain.datasource.LocalSource
-import com.gmail.laktionov.lyricsfinder.domain.datasource.LyricRepository
-import com.gmail.laktionov.lyricsfinder.domain.datasource.RemoteSource
-import com.gmail.laktionov.lyricsfinder.domain.datasource.Repository
-import com.gmail.laktionov.lyricsfinder.domain.datasource.local.DataBase
-import com.gmail.laktionov.lyricsfinder.domain.datasource.local.LocalDataSource
-import com.gmail.laktionov.lyricsfinder.domain.datasource.remote.LyricApi
-import com.gmail.laktionov.lyricsfinder.domain.datasource.remote.RemoteDataSource
+import com.gmail.laktionov.lyricsfinder.data.LyricRepository
+import com.gmail.laktionov.lyricsfinder.data.local.DataBase
+import com.gmail.laktionov.lyricsfinder.data.local.LocalDataSource
+import com.gmail.laktionov.lyricsfinder.data.local.mapper.EntityMapper
+import com.gmail.laktionov.lyricsfinder.data.remote.LyricApi
+import com.gmail.laktionov.lyricsfinder.data.remote.RemoteDataSource
+import com.gmail.laktionov.lyricsfinder.data.remote.mapper.ResponseMapper
+import com.gmail.laktionov.lyricsfinder.domain.LocalSource
+import com.gmail.laktionov.lyricsfinder.domain.RemoteSource
+import com.gmail.laktionov.lyricsfinder.domain.Repository
 import dagger.Module
 import dagger.Provides
 
@@ -27,10 +29,16 @@ object DataModule {
     )
 
     @Provides
-    fun providesLocalSource(db: DataBase): LocalSource =
-        LocalDataSource(db)
+    fun providesLocalSource(db: DataBase, mapper: EntityMapper): LocalSource =
+        LocalDataSource(db = db, mapper = mapper)
 
     @Provides
-    fun providesRemoteSource(api: LyricApi): RemoteSource =
-        RemoteDataSource(serverApi = api)
+    fun providesRemoteSource(api: LyricApi, mapper: ResponseMapper): RemoteSource =
+        RemoteDataSource(serverApi = api, mapper = mapper)
+
+    @Provides
+    fun provideEntityMapper(): EntityMapper = EntityMapper()
+
+    @Provides
+    fun provideResponseMapper(): ResponseMapper = ResponseMapper()
 }
