+++ app/src/main/java/com/gmail/laktionov/lyricsfinder/domain/model/SongLyric.kt
@@ -1,0 +1,9 @@
+package com.gmail.laktionov.lyricsfinder.domain.model
+
+data class SongLyric(
+    val artistName: String = "",
+    val songName: String = "",
+    val songText: String = ""
+) {
+    fun isEmpty() = songText.isEmpty()
+}
