+++ app/src/main/java/com/gmail/laktionov/lyricsfinder/ui/ViewModelFactory.kt
@@ -2,7 +2,7 @@ package com.gmail.laktionov.lyricsfinder.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.gmail.laktionov.lyricsfinder.domain.datasource.Repository
+import com.gmail.laktionov.lyricsfinder.domain.Repository
 import com.gmail.laktionov.lyricsfinder.ui.search.SearchViewModel
 import javax.inject.Inject
