+++ app/src/main/java/com/gmail/laktionov/lyricsfinder/ui/search/SearchFragment.kt
@@ -7,8 +7,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.gmail.laktionov.lyricsfinder.R
 import com.gmail.laktionov.lyricsfinder.appComponent
-import com.gmail.laktionov.lyricsfinder.domain.hideKeyboard
 import com.gmail.laktionov.lyricsfinder.ui.ViewModelFactory
+import com.gmail.laktionov.lyricsfinder.ui.hideKeyboard
 import kotlinx.android.synthetic.main.fragment_search.*
 import javax.inject.Inject
